Update index safely between threads. Fixes #160 and #326.
As mentioned in a warning, as well as https://github.com/fluent/fluent-plugin-s3/issues/326
and https://github.com/fluent/fluent-plugin-s3/issues/160, the process of
determining the index added to the default object key is not thread-safe. This
adds some thread-safety until version 2.x is out where chunk_id is used
instead of an index value.

This is not a perfect implementation, since there can still be races between
different workers if workers are enabled in fluentd, or if there are multiple
fluentd instances uploading to the same bucket. This commit is just to resolve
this problem short-term in a way that's backwards compatible.

Signed-off-by: William Orr <will@worrbase.com>